requirements.txt
fastapi
uvicorn[standard]
pymongo
python-multipart
cryptography
authlib
httpx
itsdangerous
python-dotenv
watchfiles
fhir.resources
